[#2] :truck: rename(icon): icon 폴더 정리
