+++ client/src/assets/icons/diary/notebook_love.png
