+++ client/src/assets/icons/nav-bar/colored_grape.png
