+++ client/src/assets/icons/nav-bar/colored_note.png
