+++ client/src/assets/icons/nav-bar/gray_grape.png
