+++ client/src/assets/icons/nav-bar/gray_note.png
